separate tabs into components
