+++ src/workspaces/NewWorkspaceModal/NewWorkspaceWizard.tsx
@@ -1,53 +1,31 @@
-import {
-  ButtonSecondary,
-  ExternalLink,
-  Icon,
-  Modal,
-  modalStyles,
-  Switch,
-  TooltipTrigger,
-} from '@terra-ui-packages/components';
+import { ButtonSecondary, Icon, Modal, modalStyles } from '@terra-ui-packages/components';
 import _ from 'lodash/fp';
 import React, { ReactNode, useState } from 'react';
 import { defaultLocation } from 'src/analysis/utils/runtime-utils';
-import { validateUserEmails } from 'src/billing/utils';
 import { BillingProject, CloudPlatform, GCPBillingProject } from 'src/billing-core/models';
-import { CloudProviderIcon } from 'src/components/CloudProviderIcon';
-import { ButtonPrimary, IdContainer, Link, Select, spinnerOverlay, VirtualizedSelect } from 'src/components/common';
-import { InfoBox } from 'src/components/InfoBox';
-import { TextArea, ValidatedInput } from 'src/components/input';
-import { allRegions, availableBucketRegions, isSupportedBucketLocation } from 'src/components/region-common';
+import { ButtonPrimary, spinnerOverlay } from 'src/components/common';
+import { isSupportedBucketLocation } from 'src/components/region-common';
 import { TabWizard } from 'src/components/tabWizard';
-import { EmailSelect } from 'src/groups/Members/EmailSelect';
 import { Billing } from 'src/libs/ajax/billing/Billing';
 import { FirecloudBucket } from 'src/libs/ajax/firecloud/FirecloudBucket';
 import { CurrentUserGroupMembership, Groups } from 'src/libs/ajax/Groups';
 import { Metrics } from 'src/libs/ajax/Metrics';
 import { Workspaces } from 'src/libs/ajax/workspaces/Workspaces';
 import colors from 'src/libs/colors';
-import { getConfig } from 'src/libs/config';
-import { reportErrorAndRethrow, withErrorReportingInModal } from 'src/libs/error';
+import { withErrorReportingInModal } from 'src/libs/error';
 import Events, { extractCrossWorkspaceDetails, extractWorkspaceDetails } from 'src/libs/events';
-import { FormLabel } from 'src/libs/forms';
 import * as Nav from 'src/libs/nav';
 import { useCancellation, useOnMount, withDisplayName } from 'src/libs/react-utils';
-import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
-import { append, summarizeErrors } from 'src/libs/utils';
 import { CloneEgressWarning } from 'src/workspaces/NewWorkspaceModal/CloneEgressWarning';
 import { CreatingWorkspaceMessage } from 'src/workspaces/NewWorkspaceModal/CreatingWorkspaceMessage';
-import {
-  cloudProviderLabels,
-  isGoogleWorkspace,
-  isProtectedWorkspace,
-  WorkspaceInfo,
-  WorkspaceWrapper,
-} from 'src/workspaces/utils';
+import { isGoogleWorkspace, isProtectedWorkspace, WorkspaceInfo, WorkspaceWrapper } from 'src/workspaces/utils';
 import validate from 'validate.js';
 
-import { AccessEntry, WorkspaceAcl } from '../acl-utils';
-import { AclInput } from '../ShareWorkspaceModal/Collaborator';
-import { CurrentCollaborators } from '../ShareWorkspaceModal/CurrentCollaborators';
+import { WorkspaceAcl } from '../acl-utils';
+import { NewWorkspaceBasicTab } from './NewWorkspaceBasicTab';
+import { NewWorkspaceSecurityTab } from './NewWorkspaceSecurityTab';
+import { NewWorkspaceSharingTab } from './NewWorkspaceSharingTab';
 
 const constraints = {
   name: {
@@ -61,13 +39,6 @@ const constraints = {
   namespace: {
     presence: true,
   },
-};
-
-const invalidBillingAccountMsg =
-  'Workspaces may only be created in billing projects that have a Google billing account accessible in Terra';
-
-const ariaInvalidBillingAccountMsg = (invalidBillingAccount: boolean): string => {
-  return invalidBillingAccount ? ` with warning "${invalidBillingAccountMsg}"` : '';
 };
 
 export interface NewWorkspaceWizardProps {
@@ -108,15 +79,14 @@ export const NewWorkspaceWizard = withDisplayName(
     const [description, setDescription] = useState(cloneWorkspace?.workspace.attributes?.description || '');
     const [groups, setGroups] = useState<string[]>([]);
     const [enhancedBucketLogging, setEnhancedBucketLogging] = useState(!!requireEnhancedBucketLogging);
-    const [nameModified, setNameModified] = useState(false);
     const [loading, setLoading] = useState(true);
     const [creating, setCreating] = useState(false);
     const [createError, setCreateError] = useState<string>();
     const [bucketLocation, setBucketLocation] = useState(defaultLocation);
     const [sourceGCPWorkspaceRegion, setSourceGCPWorkspaceRegion] = useState<string>(defaultLocation);
     const [sourceGCPWorkspaceRegionError, setSourceGCPWorkspaceRegionError] = useState(false);
-    const [isAlphaRegionalityUser, setIsAlphaRegionalityUser] = useState(false);
     const [activeTab, setActiveTab] = useState<string>('basic');
+    const [acl, setAcl] = useState<WorkspaceAcl>([]);
     const signal = useCancellation();
 
     // Helpers
@@ -125,10 +95,6 @@ export const NewWorkspaceWizard = withDisplayName(
         ...(cloneWorkspace ? _.map('membersGroupName', cloneWorkspace.workspace.authorizationDomain) : []),
         ...(requiredAuthDomain ? [requiredAuthDomain] : []),
       ]);
-
-    const loadAlphaRegionalityUser = reportErrorAndRethrow('Error loading regionality group membership')(async () => {
-      setIsAlphaRegionalityUser(await Groups(signal).group(getConfig().alphaRegionalityGroup).isMember());
-    });
 
     const create = async (): Promise<void> => {
       try {
@@ -286,7 +252,6 @@ export const NewWorkspaceWizard = withDisplayName(
         setEnhancedBucketLogging(true);
       }
       loadData();
-      loadAlphaRegionalityUser();
     });
 
     // Render
@@ -295,316 +260,6 @@ export const NewWorkspaceWizard = withDisplayName(
     const errors = validate({ namespace, name }, constraints, {
       prettify: (v) => ({ namespace: 'Billing project', name: 'Name' }[v] || validate.prettify(v)),
     });
-
-    const onFocusAria = ({ focused, isDisabled }) => {
-      return `${isDisabled ? 'Disabled option ' : 'Option '}${focused['aria-label']}, focused.`;
-    };
-
-    const onChangeAria = ({ value }) => {
-      return !value ? '' : `Option ${value['aria-label']} selected.`;
-    };
-
-    const endingNotice = renderNotice ? renderNotice({ selectedBillingProject }) : undefined;
-
-    const renderPolicyAndWorkspaceInfo = () => {
-      // If we display the Azure policy/workspace section, we render the optional notice within that block
-      return endingNotice ? <div style={{ ...Style.elements.noticeContainer }}>{endingNotice}</div> : undefined;
-    };
-
-    const basicTab = (
-      <>
-        <IdContainer>
-          {(id) => (
-            <>
-              <FormLabel htmlFor={id} required>
-                Workspace name
-              </FormLabel>
-              <ValidatedInput
-                inputProps={{
-                  id,
-                  autoFocus: true,
-                  placeholder: 'Enter a name',
-                  value: name,
-                  onChange: (v) => {
-                    setName(v);
-                    setNameModified(true);
-                  },
-                }}
-                error={Utils.summarizeErrors(nameModified && errors?.name)}
-              />
-            </>
-          )}
-        </IdContainer>
-        <IdContainer>
-          {(id) => (
-            <>
-              <FormLabel htmlFor={id}>Description</FormLabel>
-              <TextArea
-                id={id}
-                style={{ height: 100 }}
-                placeholder='Enter a description'
-                value={description}
-                onChange={setDescription}
-              />
-            </>
-          )}
-        </IdContainer>
-        <IdContainer>
-          {(id) => (
-            <>
-              <FormLabel htmlFor={id} required>
-                Billing project
-              </FormLabel>
-              <VirtualizedSelect
-                id={id}
-                isClearable={false}
-                placeholder='Select a billing project'
-                value={namespace || null}
-                ariaLiveMessages={{ onFocus: onFocusAria, onChange: onChangeAria }}
-                onChange={(opt) => setNamespace(opt!.value)}
-                styles={{ option: (provided) => ({ ...provided, padding: 10 }) }}
-                options={_.map((project: BillingProject) => {
-                  const { projectName, invalidBillingAccount, cloudPlatform } = project;
-                  return {
-                    'aria-label': `${cloudProviderLabels[cloudPlatform]} ${projectName}${ariaInvalidBillingAccountMsg(
-                      invalidBillingAccount
-                    )}`,
-                    label: (
-                      <TooltipTrigger content={invalidBillingAccount && invalidBillingAccountMsg} side='left'>
-                        <div style={{ display: 'flex', alignItems: 'center' }}>
-                          {cloudPlatform === 'GCP' && (
-                            <CloudProviderIcon
-                              key={projectName}
-                              cloudProvider={cloudPlatform}
-                              style={{ marginRight: '0.5rem' }}
-                            />
-                          )}
-                          {projectName}
-                        </div>
-                      </TooltipTrigger>
-                    ),
-                    value: projectName,
-                    isDisabled: invalidBillingAccount,
-                  };
-                }, _.sortBy('projectName', _.filter({ cloudPlatform: 'GCP' }, billingProjects)))}
-              />
-            </>
-          )}
-        </IdContainer>
-        {isGoogleBillingProject() && (
-          <IdContainer>
-            {(id) => (
-              <>
-                <FormLabel htmlFor={id}>
-                  Bucket location
-                  <InfoBox style={{ marginLeft: '0.25rem' }}>
-                    <p style={{ marginTop: '0.2rem' }}>
-                      By default, workflow and Cloud Environments will run in the same region as the workspace bucket.
-                      Changing bucket or Cloud Environment locations from the defaults can lead to network egress
-                      charges.
-                    </p>
-                    <Link href='https://support.terra.bio/hc/en-us/articles/360058964552' {...Utils.newTabLinkProps}>
-                      Read more about bucket locations
-                    </Link>
-                  </InfoBox>
-                </FormLabel>
-                <Select<string>
-                  isDisabled
-                  id={id}
-                  value={bucketLocation}
-                  onChange={(opt) => setBucketLocation(opt!.value)}
-                  options={isAlphaRegionalityUser ? allRegions : availableBucketRegions}
-                />
-              </>
-            )}
-          </IdContainer>
-        )}
-      </>
-    );
-
-    /** ******************** */
-    // TODO more intelligently share this with ShareWorkspaceModal.tsx
-
-    const defaultAcl: AccessEntry = {
-      email: '',
-      accessLevel: 'READER',
-      pending: false,
-      canShare: false,
-      canCompute: false,
-    };
-
-    // State
-    const [searchValues, setSearchValues] = useState<string[]>([]);
-    const [acl, setAcl] = useState<WorkspaceAcl>([]);
-    const [newAcl, setNewAcl] = useState<AccessEntry>(defaultAcl);
-    const [lastAddedEmail, setLastAddedEmail] = useState<string | undefined>(undefined);
-
-    // // Render
-    const sharingErrors = validateUserEmails(searchValues);
-    const aclEmails = _.map('email', acl);
-
-    const addCollaborators = (collaboratorEmails: string[], collaboratorAcl: AccessEntry) => {
-      collaboratorEmails.forEach((collaboratorEmail: string) => {
-        if (!validate.single(collaboratorEmail, { email: true, exclusion: aclEmails })) {
-          setAcl(append({ ...collaboratorAcl, email: collaboratorEmail } as AccessEntry));
-          setLastAddedEmail(collaboratorEmail);
-        }
-      });
-      // Clear the search values and new acl after adding collaborators
-      setSearchValues([]);
-      setNewAcl(defaultAcl);
-    };
-
-    const sharingTab = (
-      <>
-        <div style={{ display: 'flex', alignItems: 'flex-end', gap: '0.5rem', marginTop: '1rem' }}>
-          <div style={{ flexGrow: 100, alignSelf: 'flex-start' }}>
-            <EmailSelect
-              placeholder='Add people or groups'
-              options={[]}
-              emails={searchValues}
-              setEmails={setSearchValues}
-            />
-          </div>
-
-          <div style={{ flexGrow: 1, alignSelf: 'flex-start' }}>
-            <FormLabel id='aclInputId' style={{ marginTop: '0.25rem' }}>
-              Permissions
-            </FormLabel>
-            <AclInput
-              aria-label='permissions for new collaborator'
-              value={newAcl}
-              onChange={setNewAcl}
-              disabled={false}
-              maxAccessLevel='OWNER' // presumably since we're creating the workspace
-              isAzureWorkspace={false}
-              showRow={false}
-            />
-          </div>
-          <div style={{ flexGrow: 1, alignSelf: 'flex-start', marginTop: '1.65rem' }}>
-            <ButtonPrimary
-              disabled={!!sharingErrors}
-              tooltip={summarizeErrors(sharingErrors)}
-              onClick={() => addCollaborators(searchValues, newAcl)}
-            >
-              Add
-            </ButtonPrimary>
-          </div>
-        </div>
-        <CurrentCollaborators
-          acl={acl}
-          setAcl={setAcl}
-          originalAcl={acl}
-          lastAddedEmail={lastAddedEmail}
-          workspaceAccessLevel='OWNER'
-          isAzureWorkspace={false}
-          listStyles={{ margin: '0.5rem 0rem 0px' }}
-        />
-      </>
-    );
-
-    /** *************************** */
-
-    const securityTab = (
-      <div style={{ margin: '1rem 0.25rem 0.25rem 0' }}>
-        <IdContainer>
-          {(id) => (
-            <>
-              <FormLabel htmlFor={id}>Additional Security Options</FormLabel>
-              <p style={{ marginTop: '.25rem', marginLeft: '0.25rem' }}>
-                In order to ensure that only authorized Terra Users get access to a User’s Workspace you must select
-                these options:
-              </p>
-            </>
-          )}
-        </IdContainer>
-        <IdContainer>
-          {(id) => (
-            <div
-              style={{
-                display: 'flex',
-                flexDirection: 'row',
-                outline: 'solid',
-                borderRadius: '5px',
-                columnGap: '5%',
-              }}
-            >
-              <div
-                style={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  width: '80%',
-                  justifyContent: 'space-between',
-                  padding: '1rem',
-                }}
-              >
-                <FormLabel htmlFor={id}>Enable Secure Monitoring</FormLabel>
-                <p style={{ padding: '0.1rem' }}>
-                  Enhanced logging and monitoring are enabled to support the use of controlled-access data in this
-                  workspace.
-                </p>
-              </div>
-              <div style={{ marginTop: '0.5rem', alignSelf: 'right', display: 'flex', alignItems: 'center' }}>
-                <Switch
-                  onLabel=''
-                  offLabel=''
-                  onChange={() => {
-                    setEnhancedBucketLogging(!enhancedBucketLogging);
-                  }}
-                  checked={enhancedBucketLogging || groups.length > 0}
-                  width={40}
-                  height={20}
-                  disabled={!!requireEnhancedBucketLogging || groups.length > 0 || cloningGcpProtectedWorkspace}
-                  aria-describedby='Enable additional security monitoring'
-                />
-              </div>
-            </div>
-          )}
-        </IdContainer>
-        <IdContainer>
-          {(id) => (
-            <>
-              <FormLabel htmlFor={id}>Authorization Domain (optional)</FormLabel>
-              <div style={{ marginLeft: '0.25rem' }}>
-                Authorization Domains restrict data access to only specified individuals in a group and are intended to
-                fulfill requirements you may have for data governed by a compliance standard, such as federal
-                controlled-access data or HIPAA protected data. They follow all workspace copies and cannot be removed.
-                <p>
-                  For more details, see{' '}
-                  <ExternalLink
-                    href='https://support.terra.bio/hc/en-us/articles/360026775691'
-                    {...Utils.newTabLinkProps}
-                  >
-                    When to use an Authorization Domain
-                  </ExternalLink>
-                  .
-                </p>
-              </div>
-              {!!existingGroups.length && (
-                <div style={{ marginBottom: '0.5rem', fontSize: 12 }}>
-                  <div style={{ marginBottom: '0.2rem' }}>Inherited groups:</div>
-                  {existingGroups.join(', ')}
-                </div>
-              )}
-              <Select<string, true>
-                id={id}
-                isClearable={false}
-                isMulti
-                placeholder='Select groups'
-                isDisabled={!allGroups || !billingProjects}
-                value={groups}
-                onChange={(data) => {
-                  setGroups(_.map('value', data));
-                  setEnhancedBucketLogging(!!requireEnhancedBucketLogging || data.length > 0);
-                }}
-                options={_.difference(_.uniq(_.map('groupName', allGroups)), existingGroups).sort()}
-              />
-            </>
-          )}
-        </IdContainer>
-        {renderPolicyAndWorkspaceInfo()}
-      </div>
-    );
 
     const handleSecondaryButtonClick = async () =>
       Utils.switchCase(
@@ -741,12 +396,12 @@ export const NewWorkspaceWizard = withDisplayName(
                   aria-label='create workspace wizard'
                   activeTab={activeTab}
                   tabs={[
-                    { name: 'basic', displayName: '1. Basic Information and Billing', disabled: false },
-                    { name: 'sharing', displayName: '2. Sharing', disabled: activeTab === 'basic' && !!errors },
+                    { name: 'basic', displayName: '1. Basic Information and Billing', disabled: () => false },
+                    { name: 'sharing', displayName: '2. Sharing', disabled: () => activeTab === 'basic' && !!errors },
                     {
                       name: 'security',
                       displayName: '3. Additional Security Options',
-                      disabled: activeTab === 'basic' && !!errors,
+                      disabled: () => activeTab === 'basic' && !!errors,
                     },
                   ]}
                   style={{ textTransform: 'none', width: '90%' }}
@@ -757,9 +412,36 @@ export const NewWorkspaceWizard = withDisplayName(
                 >
                   {null /* nothing to display at the end of the tab bar */}
                 </TabWizard>
-                {activeTab === 'basic' && basicTab}
-                {activeTab === 'sharing' && sharingTab}
-                {activeTab === 'security' && securityTab}
+                {activeTab === 'basic' && (
+                  <NewWorkspaceBasicTab
+                    setName={setName}
+                    name={name}
+                    errors={errors}
+                    description={description}
+                    setDescription={setDescription}
+                    namespace={namespace}
+                    billingProjects={billingProjects}
+                    setNamespace={setNamespace}
+                    bucketLocation={bucketLocation}
+                    setBucketLocation={setBucketLocation}
+                  />
+                )}
+                {activeTab === 'sharing' && <NewWorkspaceSharingTab acl={acl} setAcl={setAcl} />}
+                {activeTab === 'security' && (
+                  <NewWorkspaceSecurityTab
+                    requireEnhancedBucketLogging={requireEnhancedBucketLogging}
+                    setEnhancedBucketLogging={setEnhancedBucketLogging}
+                    groups={groups}
+                    setGroups={setGroups}
+                    cloningGcpProtectedWorkspace={cloningGcpProtectedWorkspace}
+                    existingGroups={existingGroups}
+                    allGroups={allGroups}
+                    billingProjects={billingProjects}
+                    renderNotice={renderNotice}
+                    selectedBillingProject={selectedBillingProject}
+                    enhancedBucketLogging={enhancedBucketLogging}
+                  />
+                )}
 
                 {!!selectedBillingProject && !!cloneWorkspace && activeTab === 'basic' && (
                   <CloneEgressWarning
